feat: extract toast limit constant
