+++ src/constants/index.ts
@@ -1,2 +1,3 @@
 export { API, API_URL } from './apiUrl';
 export { messageCodeMap } from './messageCodeMap';
+export { MAX_VISIBLE_TOASTS } from './toast';
