+++ src/constants/toast.ts
@@ -1,0 +1,1 @@
+export const MAX_VISIBLE_TOASTS = 5;
